revisão de comentários
