--- app/src/main/java/com/example/challengecskotlin/ApiService.kt
@@ -1,20 +1,0 @@
-package com.example.challengecskotlin
-
-import com.example.challengecskotlin.GithubApi.PARAM_PAGE
-import com.example.challengecskotlin.GithubApi.PARAM_QUERY
-import com.example.challengecskotlin.GithubApi.PARAM_SORT
-import retrofit2.Call
-import retrofit2.http.GET
-import retrofit2.http.Path
-import retrofit2.http.Query
-
-interface ApiService {
-    @GET("search/repositories")
-    fun fetchAllUsers(@Query(PARAM_QUERY) query: String,
-                      @Query(PARAM_SORT) sort: String,
-                      @Query(PARAM_PAGE) page: String) : Call<SearchResponse>
-
-    @GET("/repos/{login}/{name}/pulls")
-    fun fetchPullRequests(@Path("login") login: String,
-                          @Path("name") name: String) : Call<List<PullRequestObject>>
-}
