--- app/src/main/java/com/example/challengecskotlin/SearchResponse.kt
@@ -1,9 +1,0 @@
-package com.example.challengecskotlin
-
-class SearchResponse {
-    val items: List<Repo> = ArrayList<Repo>()
-
-    fun getRepos(): List<Repo> {
-        return items
-    }
-}
