+++ app/src/main/java/com/example/challengecskotlin/activities/MainActivity.kt
@@ -1,4 +1,4 @@
-package com.example.challengecskotlin
+package com.example.challengecskotlin.activities
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
@@ -13,6 +13,12 @@ import retrofit2.Call
 import retrofit2.Callback
 import retrofit2.Response
 import androidx.recyclerview.widget.DividerItemDecoration
+import com.example.challengecskotlin.utils.PaginationScrollListener
+import com.example.challengecskotlin.R
+import com.example.challengecskotlin.adapters.RepoAdapter
+import com.example.challengecskotlin.api.GithubApi
+import com.example.challengecskotlin.models.Repo
+import com.example.challengecskotlin.models.SearchResponse
 
 
 class MainActivity : AppCompatActivity() {
@@ -36,11 +42,8 @@ class MainActivity : AppCompatActivity() {
         loadingBar = findViewById(R.id.main_progress)
         val recyclerView = findViewById<RecyclerView>(R.id.recyclerView)
         val linearLayoutManager = LinearLayoutManager(this)
+        val dividerItemDecoration = DividerItemDecoration(recyclerView.context, linearLayoutManager.orientation)
 
-        val dividerItemDecoration = DividerItemDecoration(
-            recyclerView.context,
-            linearLayoutManager.orientation
-        )
         recyclerView.addItemDecoration(dividerItemDecoration)
         recyclerView.layoutManager = linearLayoutManager
         recyclerView.itemAnimator = DefaultItemAnimator()
@@ -62,11 +65,9 @@ class MainActivity : AppCompatActivity() {
         loadFirstPage()
 
     }
-
+    //carregamento da primeira requisição à API
     private fun loadFirstPage() {
-        //d(TAG, "loadFirstPage: $PAGE_START")
-
-        GithubApi.searchService.fetchAllUsers(QUERY, SORT, PAGE_START.toString()).enqueue(object : Callback<SearchResponse> {
+        GithubApi.searchService.fetchRepositories(QUERY, SORT, PAGE_START.toString()).enqueue(object : Callback<SearchResponse> {
             override fun onResponse(call: Call<SearchResponse>, response: Response<SearchResponse>) {
                 if(response.isSuccessful){
                     //Enviando os dados pro adapter
@@ -75,6 +76,7 @@ class MainActivity : AppCompatActivity() {
                     adapter.addAll(results)
 
                 }
+                //verificando se há mais páginas (paginação)
                 if (currentPage <= TOTAL_PAGES)
                     adapter.addLoadingFooter()
                 else
@@ -82,22 +84,21 @@ class MainActivity : AppCompatActivity() {
             }
 
             override fun onFailure(call: Call<SearchResponse>, t: Throwable) {
-                //d("onFailure", "fail:" + t.message)
                 loadingBar.visibility = View.GONE
                 Toast.makeText(this@MainActivity, "Erro: ${t.message}", Toast.LENGTH_LONG).show()
             }
         })
     }
 
+    //pega a resposta da API
     private fun fetchResults(response: Response<SearchResponse>): List<Repo> {
         val repositories = response.body()
         return repositories!!.getRepos()
     }
 
+    //carregamento das páginas posteriores a primeira
     private fun loadNextPage() {
-        //d(TAG, "loadNextPage: $currentPage")
-
-        GithubApi.searchService.fetchAllUsers(QUERY, SORT, currentPage.toString()).enqueue(object : Callback<SearchResponse> {
+        GithubApi.searchService.fetchRepositories(QUERY, SORT, currentPage.toString()).enqueue(object : Callback<SearchResponse> {
             override fun onResponse(call: Call<SearchResponse>, response: Response<SearchResponse>) {
                 adapter.removeLoadingFooter()
                 isLoading = false
