+++ app/src/main/java/com/example/challengecskotlin/adapters/RepoAdapter.kt
@@ -1,4 +1,4 @@
-package com.example.challengecskotlin
+package com.example.challengecskotlin.adapters
 
 import android.content.Context
 import android.content.Intent
@@ -7,10 +7,12 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import android.widget.TextView
-import androidx.core.content.ContextCompat.startActivity
 import androidx.recyclerview.widget.RecyclerView
 import androidx.recyclerview.widget.RecyclerView.ViewHolder
 import com.bumptech.glide.Glide
+import com.example.challengecskotlin.activities.PullRequestActivity
+import com.example.challengecskotlin.R
+import com.example.challengecskotlin.models.Repo
 import de.hdodenhof.circleimageview.CircleImageView
 
 
@@ -28,6 +30,7 @@ open class RepoAdapter(context: Context) : RecyclerView.Adapter<ViewHolder> () {
         var viewHolder: ViewHolder? = null
         val inflater = LayoutInflater.from(parent.context)
 
+        //verificando o tipo de view: conteúdo ou carregamento
         when (viewType) {
             ITEM -> viewHolder = getViewHolder(parent, inflater)
             LOADING -> {
@@ -55,13 +58,11 @@ open class RepoAdapter(context: Context) : RecyclerView.Adapter<ViewHolder> () {
             ITEM -> {
                 val repoVH = holder as RepoVH
                 repoVH.name.text = repo.name
-                //repoVH.description.text = repo.description
                 repoVH.login.text = repo.description
                 repoVH.forks.text = repo.forks
                 repoVH.stars.text = repo.stargazers_count
                 repoVH.username.text = repo.owner!!.login
-                holder.itemView.setOnClickListener {
-                    d("onClick", "clicado: $repo")
+                holder.itemView.setOnClickListener { //passando dados pra pull request acitivity via onClick
                     val intent = Intent(mContext, PullRequestActivity::class.java)
                     intent.putExtra("login", repo.owner.login)
                     intent.putExtra("repositoryName", repo.name)
@@ -75,8 +76,8 @@ open class RepoAdapter(context: Context) : RecyclerView.Adapter<ViewHolder> () {
 
             }
             LOADING -> {
-            }
-        }// Do nothing
+            } //carrega progress bar
+        }
     }
 
     override fun getItemViewType(position: Int): Int {
@@ -95,32 +96,30 @@ open class RepoAdapter(context: Context) : RecyclerView.Adapter<ViewHolder> () {
 
     protected inner class LoadingVH(itemView: View) : RecyclerView.ViewHolder(itemView)
 
+    //adiciona item na lista repoResults
     private fun add(r: Repo) {
         repoResults.add(r)
         notifyItemInserted(repoResults.size - 1)
     }
 
+    //adiciona todos os itens a lista de repositórios
     fun addAll(repoResults: List<Repo>) {
         for (result in repoResults) {
             add(result)
         }
     }
-    
 
+    //adiciona a progress bar
     fun addLoadingFooter() {
         isLoadingAdded = true
         add(Repo())
     }
 
+    //remove progress bar
     fun removeLoadingFooter() {
         isLoadingAdded = false
         val position = repoResults.size - 1
-        //val result = getItem(position)
         repoResults.removeAt(position)
         notifyItemRemoved(position)
     }
-
-//    private fun getItem(position: Int): Repo {
-//        return repoResults[position]
-//    }
 }
