+++ app/src/main/java/com/example/challengecskotlin/models/PullRequestObject.kt
@@ -1,4 +1,4 @@
-package com.example.challengecskotlin
+package com.example.challengecskotlin.models
 
 class PullRequestObject {
     val title: String = ""
