+++ app/src/main/java/com/example/challengecskotlin/models/SearchResponse.kt
@@ -1,0 +1,11 @@
+package com.example.challengecskotlin.models
+
+import com.example.challengecskotlin.models.Repo
+
+class SearchResponse {
+    val items: List<Repo> = ArrayList()
+
+    fun getRepos(): List<Repo> {
+        return items
+    }
+}
