+++ app/src/main/res/layout/activity_pull_request.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".PullRequestActivity">
+    tools:context=".activities.PullRequestActivity">
 
     <androidx.appcompat.widget.Toolbar
         android:id="@+id/my_toolbar"
